requirements.txt
pytest==7.4.4
selenium==4.15.2
requests==2.31.0
allure-pytest==2.13.0
python-dotenv==1.0.0
pydantic==2.5.0
webdriver-manager==4.0.1